+++ changelog.md
@@ -2,6 +2,7 @@
 
 * Teach `StagedFile`s how to calculate which lines were actually added/modified
 * Checks no longer need to filter temporary staged file paths themselves
+* Condense Ruby style check output
 
 0.1.11